refactor(titiler-cog): address code review
- Extract titiler helpers (URLs, buildDescriptor, decode, getTileData,
  renderTile) into src/titiler.ts. App.tsx now only handles the React
  UI and layer wiring.
- Drop unused @developmentseed/proj dep; add peer dep @deck.gl/mesh-layers.
- Replace `parsed.data as Uint8Array` cast with an instanceof guard.
- Note texture-destroy semantics in getTileData docstring.
