+++ examples/titiler-cog/package.json
@@ -14,9 +14,9 @@
     "@deck.gl/geo-layers": "^9.3.0",
     "@deck.gl/layers": "^9.3.0",
     "@deck.gl/mapbox": "^9.3.0",
+    "@deck.gl/mesh-layers": "^9.3.0",
     "@developmentseed/deck.gl-raster": "workspace:^",
     "@developmentseed/morecantile": "workspace:^",
-    "@developmentseed/proj": "workspace:^",
     "@luma.gl/core": "9.3.2",
     "maplibre-gl": "^5.19.0",
     "npyjs": "^1.1.0",
