+++ examples/titiler-cog/src/App.tsx
@@ -1,157 +1,25 @@
-import type { _TileLoadProps as TileLoadProps } from "@deck.gl/geo-layers";
 import type { MapboxOverlayProps } from "@deck.gl/mapbox";
 import { MapboxOverlay } from "@deck.gl/mapbox";
-import type {
-  GetTileDataOptions,
-  MinimalTileData,
-  RasterModule,
-  RenderTileResult,
-  TilesetDescriptor,
-} from "@developmentseed/deck.gl-raster";
-import {
-  RasterTileLayer,
-  TileMatrixSetAdaptor,
-} from "@developmentseed/deck.gl-raster";
-import {
-  CreateTexture,
-  MaskTexture,
-} from "@developmentseed/deck.gl-raster/gpu-modules";
+import type { TilesetDescriptor } from "@developmentseed/deck.gl-raster";
+import { RasterTileLayer } from "@developmentseed/deck.gl-raster";
 import type { TileMatrixSet } from "@developmentseed/morecantile";
-import type { Texture } from "@luma.gl/core";
 import "maplibre-gl/dist/maplibre-gl.css";
-import npyjs from "npyjs";
-import proj4 from "proj4";
 import { useEffect, useRef, useState } from "react";
 import type { MapRef } from "react-map-gl/maplibre";
 import { Map as MaplibreMap, useControl } from "react-map-gl/maplibre";
+import type { InfoResponse, TileData } from "./titiler";
+import {
+  buildDescriptor,
+  COG_URL,
+  getTileData,
+  renderTile,
+  TITILER_BASE,
+} from "./titiler";
 
 function DeckGLOverlay(props: MapboxOverlayProps) {
   const overlay = useControl<MapboxOverlay>(() => new MapboxOverlay(props));
   overlay.setProps(props);
   return null;
-}
-
-const COG_URL =
-  "https://sentinel-cogs.s3.us-west-2.amazonaws.com/sentinel-s2-l2a-cogs/18/T/WL/2026/1/S2B_18TWL_20260101_0_L2A/TCI.tif";
-const TITILER_BASE = "https://titiler.xyz";
-
-type InfoResponse = {
-  bounds: [number, number, number, number]; // WGS84 [w, s, e, n]
-  band_descriptions?: [string, Record<string, unknown>][];
-  dtype?: string;
-  [key: string]: unknown;
-};
-
-function buildDescriptor(tms: TileMatrixSet): TilesetDescriptor {
-  const converter = proj4("EPSG:3857", "EPSG:4326");
-  const projectTo4326 = (x: number, y: number) =>
-    converter.forward<[number, number]>([x, y], false);
-  const projectFrom4326 = (x: number, y: number) =>
-    converter.inverse<[number, number]>([x, y], false);
-  const identity = (x: number, y: number): [number, number] => [x, y];
-  return new TileMatrixSetAdaptor(tms, {
-    projectTo3857: identity,
-    projectFrom3857: identity,
-    projectTo4326,
-    projectFrom4326,
-  });
-}
-
-type TileData = MinimalTileData & {
-  texture: Texture;
-  mask?: Texture;
-};
-
-/**
- * Repack a band-separate uint8 buffer of shape [B, H, W] into an
- * interleaved RGBA uint8 buffer of length H*W*4. Bands 0-2 go to R/G/B;
- * alpha is always 255 (the 4th titiler band is a mask, handled separately).
- */
-function repackToRGBA(
-  bandSeparate: Uint8Array,
-  height: number,
-  width: number,
-): Uint8Array {
-  const pixelCount = height * width;
-  const rgba = new Uint8Array(pixelCount * 4);
-  const bandOffset0 = 0;
-  const bandOffset1 = pixelCount;
-  const bandOffset2 = 2 * pixelCount;
-  for (let i = 0; i < pixelCount; i++) {
-    rgba[i * 4] = bandSeparate[bandOffset0 + i]!;
-    rgba[i * 4 + 1] = bandSeparate[bandOffset1 + i]!;
-    rgba[i * 4 + 2] = bandSeparate[bandOffset2 + i]!;
-    rgba[i * 4 + 3] = 255;
-  }
-  return rgba;
-}
-
-function tileNpyUrl(x: number, y: number, z: number): string {
-  return `${TITILER_BASE}/cog/tiles/WebMercatorQuad/${z}/${x}/${y}.npy?url=${encodeURIComponent(COG_URL)}`;
-}
-
-async function getTileData(
-  tile: TileLoadProps,
-  options: GetTileDataOptions,
-): Promise<TileData> {
-  const { device, signal } = options;
-  const { x, y, z } = tile.index;
-  const response = await fetch(tileNpyUrl(x, y, z), { signal });
-  if (!response.ok) {
-    throw new Error(
-      `titiler tile ${z}/${x}/${y} ${response.status}: ${await response.text()}`,
-    );
-  }
-  const buffer = await response.arrayBuffer();
-  const parsed = await new npyjs().load(buffer);
-  if (parsed.dtype !== "u1") {
-    throw new Error(`Expected uint8 (u1) npy, got dtype=${parsed.dtype}`);
-  }
-  if (parsed.shape.length !== 3) {
-    throw new Error(
-      `Expected shape [B, H, W], got [${parsed.shape.join(", ")}]`,
-    );
-  }
-  const [bands, height, width] = parsed.shape as [number, number, number];
-  if (bands !== 3 && bands !== 4) {
-    throw new Error(`Expected 3 or 4 bands, got ${bands}`);
-  }
-  const data = parsed.data as Uint8Array;
-  const rgba = repackToRGBA(data, height, width);
-  const texture = device.createTexture({
-    data: rgba,
-    format: "rgba8unorm",
-    width,
-    height,
-    sampler: { minFilter: "linear", magFilter: "linear" },
-  });
-  let mask: Texture | undefined;
-  let byteLength = rgba.byteLength;
-  if (bands === 4) {
-    const maskBand = data.subarray(3 * height * width, 4 * height * width);
-    mask = device.createTexture({
-      data: maskBand,
-      format: "r8unorm",
-      width,
-      height,
-      sampler: { minFilter: "nearest", magFilter: "nearest" },
-    });
-    byteLength += maskBand.byteLength;
-  }
-  return { width, height, byteLength, texture, mask };
-}
-
-function renderTile(data: TileData): RenderTileResult {
-  const pipeline: RasterModule[] = [
-    { module: CreateTexture, props: { textureName: data.texture } },
-  ];
-  if (data.mask) {
-    pipeline.push({
-      module: MaskTexture,
-      props: { maskTexture: data.mask },
-    });
-  }
-  return { renderPipeline: pipeline };
 }
 
 export default function App() {
